+++ src/features/UserProfile/components/Profile.tsx
@@ -1,24 +1,78 @@
-import * as React from "react";
+import React from "react";
+import { useIsDeveloper, useLogoutHandler } from "@/providers/Profile";
+import User from "../../../views/user/User";
+import { CookThis } from "@/features/UserProfile/components/CookThis";
+import { Developer } from "@/features/UserProfile/components/Developer";
 import type { UserType } from "@/global/types/identity";
+import { Box, Button, Grid, Paper, Stack, Typography } from "@mui/material";
+import { styled } from "@mui/material/styles";
+import Avatar from "@mui/material/Avatar";
+import { LogoutIcon } from "../../../views/common/icons";
+
+const Info = styled(Box)({
+    flex: 1,
+});
+
+const ProfileDisplay = styled(Paper)(({ theme }) => ({
+    padding: theme.spacing(2),
+}));
 
 interface Props {
-    currentUser: UserType;
+    currentUser: Pick<UserType, "imageUrl" | "name" | "email">;
 }
 
-export const Profile: React.FC<Props> = ({ currentUser }) => {
+const Profile: React.FC<Props> = ({ currentUser: user }) => {
+    const isDeveloper = useIsDeveloper();
+    const doLogout = useLogoutHandler();
+
+    const handleLogout = (e) => {
+        e.preventDefault();
+        e.stopPropagation();
+        doLogout();
+    };
+
     return (
-        <>
-            {currentUser.imageUrl && (
-                <img
-                    src={currentUser.imageUrl}
-                    alt={currentUser.name || currentUser.email || undefined}
-                    title="Holy moley, you're attractive!"
-                />
+        <Stack gap={2}>
+            <ProfileDisplay>
+                <Grid container gap={2}>
+                    <Box>
+                        {user.imageUrl && (
+                            <Avatar
+                                src={user.imageUrl}
+                                sx={{ height: 96, width: 96 }}
+                                alt={user.name || user.email || undefined}
+                                title="Holy moley, you're attractive!"
+                            />
+                        )}
+                    </Box>
+                    <Info>
+                        <Typography variant={"h2"}>{user.name}</Typography>
+                        <Typography>{user.email}</Typography>
+                    </Info>
+                    <div>
+                        <Button
+                            onClick={handleLogout}
+                            variant="outlined"
+                            startIcon={<LogoutIcon />}
+                        >
+                            Logout
+                        </Button>
+                    </div>
+                </Grid>
+            </ProfileDisplay>
+            <ProfileDisplay>
+                <CookThis />
+            </ProfileDisplay>
+            <ProfileDisplay>
+                <User {...user} />
+            </ProfileDisplay>
+            {isDeveloper && (
+                <ProfileDisplay>
+                    <Developer />
+                </ProfileDisplay>
             )}
-            <div className="profile-name">
-                <h2>{currentUser.name}</h2>
-                <p className="profile-email">{currentUser.email}</p>
-            </div>
-        </>
+        </Stack>
     );
 };
+
+export default Profile;
