--- src/features/UserProfile/data/queries.ts
@@ -1,13 +1,0 @@
-import { gql } from "@/__generated__";
-
-export const GetCurrentUserQuery = gql(`
-    query GetCurrentUser {
-        getCurrentUser {
-            id
-            name
-            email
-            imageUrl
-            provider
-        }
-    }
-`);
