--- src/views/UserProfile/UserProfileView.tsx
@@ -1,10 +1,0 @@
-import PageBody from "../common/PageBody";
-import { UserProfileController } from "@/features/UserProfile/UserProfileController";
-
-export const UserProfileView = () => {
-    return (
-        <PageBody>
-            <UserProfileController />
-        </PageBody>
-    );
-};
